Get set up with Zero 3!
Default TypeScript and Babel 7 support with Zero 3 :)
