--- .babelrc
@@ -1,1 +1,0 @@
-{"presets": ["@helpscout/zero/babel"]}
