+++ package.json
@@ -4,23 +4,18 @@
   "description": "A tiny Redux",
   "main": "dist/index.js",
   "private": false,
-  "source": "dist/index.js",
   "scripts": {
-    "check:npm": "node scripts/check-npm.js",
-    "prestart": "npm run check:npm",
-    "build:ts": "tsc",
-    "build": "npm run clean && npm run build:ts && npm run build:postclean",
+    "prestart": "zero prestart",
+    "build": "zero build && npm run build:postclean",
     "build:postclean": "rm -rf dist/__tests__ && rm -rf dist/__fixtures__",
-    "bundle": "microbundle build --output build",
+    "bundle": "zero bundle",
     "lint": "zero lint",
     "dev": "zero test --watchAll",
-    "git:push": "git push --follow-tags",
     "test": "zero test --coverage",
     "test:ci": "zero test",
     "test:update": "npm test -- --updateSnapshot --coverage",
     "validate": "zero validate",
-    "setup": "npm install && npm run validate -s",
-    "precommit": "zero precommit",
+    "precommit": "zero pre-commit",
     "clean": "rm -rf dist lib build",
     "coverage": "nyc report --temp-directory=coverage --reporter=text-lcov | coveralls",
     "release": "npm run test && npm version",
@@ -32,8 +27,8 @@
     "start": "npm run storybook",
     "storybook": "start-storybook -p 6007",
     "build-storybook": "build-storybook",
-    "prettier": "prettier \"src/**/*.{js,ts,tsx}\" --write",
-    "pretty": "npm run prettier"
+    "prettier": "zero format",
+    "pretty": "zero format"
   },
   "files": [
     "README.md",
@@ -65,25 +60,19 @@
   },
   "devDependencies": {
     "@helpscout/helix": "0.1.0",
-    "@helpscout/zero": "0.0.9",
-    "@types/enzyme": "3.1.15",
+    "@helpscout/zero": "3.0.0-0",
+    "@types/enzyme": "3.9.0",
     "@types/jest": "23.3.10",
-    "@types/react": "16.7.11",
+    "@types/react": "16.8.6",
     "@types/redux": "3.6.0",
-    "babel-jest": "23.6.0",
-    "check-dependencies": "1.1.0",
-    "coveralls": "3.0.2",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.7.0",
-    "microbundle": "0.8.1",
-    "np": "2.20.1",
-    "nyc": "13.0.0",
+    "coveralls": "3.0.3",
+    "enzyme": "3.9.0",
+    "enzyme-adapter-react-16": "1.10.0",
+    "nyc": "13.3.0",
     "react": "^16",
     "react-dom": "^16",
-    "react-scripts": "2.1.1",
     "react-test-renderer": "^16",
     "redux-thunk": "2.3.0",
-    "ts-jest": "^23.10.5",
     "typescript": "^3.2.1"
   }
 }
